+++ bin/composer.py
@@ -9,28 +9,27 @@ menu_option = [
     {"code": "update", "title": "Update", 'description': 'update dependencys'},
 ]
 
-def switchOption(code):
+def switch_option(code):
     if code in ['rmvendor', 'install','update','runscripts']: 
-        runCommandsOptions(code)
+        run_commands_options(code)
     else:
         print('You must select an option')
 
-def runCommandsOptions(action, command = ''):
-    if action == 'rmvendor' : runCommand('rm -rf vendor')
-    elif action == 'install': runCommand('composer install')
-    elif action == 'update': runCommand('composer update')
+def run_commands_options(action, command = ''):
+    if action == 'rmvendor' : run_command('rm -rf vendor')
+    elif action == 'install': run_command('composer install')
+    elif action == 'update': run_command('composer update')
     elif action == 'runscripts':
         print(common.msgColor('Composer will run followed by what you type next.','WARNING'))
         action = input(common.msgColor('Enter the action or properties: ','WARNING'))
-        runCommand(f'composer {action}')
+        run_command(f'composer {action}')
     else:
-        print('Error command.')
+        print(common.msgColor('Error command.','DANGER'))
 
-def runCommand(command):
-    common.cli(f'docker-compose run --rm app {command}', True)
+def run_command(command):
+    common.cli(f'{common.command_docker_compose()} run --rm app_cli {command}', True)
 
 if __name__ == '__main__':
-
     common.cli('clear')
     showHelper = False
     while True:
@@ -46,7 +45,7 @@ if __name__ == '__main__':
             common.cli('clear')
         elif opt.isnumeric() and (int(opt) in range(0,len(menu_option))):
             code = menu_option[int(opt)]['code']
-            switchOption(code)
+            switch_option(code)
 
             input(common.msgColor('\nPress enter to return to the menu','SUCCESS'))
             common.cli('clear')